# api.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /exercises:
    get:
      operationId: GetAllExercises
      responses:
        '200':
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseEntityObject'
    post:
      operationId: addExercise
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/ExerciseMappingHelper'
      responses:
        '200':
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseEntityObject'
  /exercises/{id}:
    put:
      operationId: updateExercise
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/ExerciseMappingHelper'
      responses:
        '200':
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseEntityObject'
  /exercise/{id}:
    get:
      operationId: getExercise
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseEntityObject'
  /users:
    get:
      operationId: GetAllUSers
      responses:
        '200':
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseEntityObject'
    post:
      operationId: addUser
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/UserMappingHelper'
      responses:
        '200':
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseEntityObject'
  /ranking:
    get:
      operationId: sortByPoints
      responses:
        '200':
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserA'
components:
  schemas:
    ContentDisposition:
      type: object
      properties:
        type:
          type: string
        name:
          type: string
        filename:
          type: string
        charset:
          type: object
          properties:
            registered:
              type: boolean
        size:
          type: integer
          format: int64
        creationDate:
          type: string
          format: date-time
        modificationDate:
          type: string
          format: date-time
        readDate:
          type: string
          format: date-time
    HttpHeaders:
      type: object
      properties:
        empty:
          type: boolean
        location:
          type: string
          format: uri
        lastModified:
          type: integer
          format: int64
        host:
          type: object
          properties:
            address:
              type: object
              properties:
                canonicalHostName:
                  type: string
                address:
                  type: array
                  items:
                    type: string
                    format: byte
                hostAddress:
                  type: string
                hostName:
                  type: string
                multicastAddress:
                  type: boolean
                anyLocalAddress:
                  type: boolean
                loopbackAddress:
                  type: boolean
                linkLocalAddress:
                  type: boolean
                siteLocalAddress:
                  type: boolean
                mcglobal:
                  type: boolean
                mcnodeLocal:
                  type: boolean
                mclinkLocal:
                  type: boolean
                mcsiteLocal:
                  type: boolean
                mcorgLocal:
                  type: boolean
            port:
              type: integer
              format: int32
            unresolved:
              type: boolean
            hostName:
              type: string
            hostString:
              type: string
        all:
          type: object
          additionalProperties:
            type: string
          writeOnly: true
        range:
          type: array
          items:
            $ref: '#/components/schemas/HttpRange'
        contentLength:
          type: integer
          format: int64
        contentType:
          $ref: '#/components/schemas/MediaType'
        date:
          type: integer
          format: int64
        ifModifiedSince:
          type: integer
          format: int64
        connection:
          type: array
          items:
            type: string
        origin:
          type: string
        acceptCharset:
          type: array
          items:
            type: object
            properties:
              registered:
                type: boolean
        allow:
          uniqueItems: true
          type: array
          items:
            type: string
            enum:
            - GET
            - HEAD
            - POST
            - PUT
            - PATCH
            - DELETE
            - OPTIONS
            - TRACE
        contentLanguage:
          type: object
          properties:
            language:
              type: string
            script:
              type: string
            country:
              type: string
            variant:
              type: string
            extensionKeys:
              uniqueItems: true
              type: array
              items:
                type: string
            unicodeLocaleAttributes:
              uniqueItems: true
              type: array
              items:
                type: string
            unicodeLocaleKeys:
              uniqueItems: true
              type: array
              items:
                type: string
            iso3Language:
              type: string
            iso3Country:
              type: string
            displayLanguage:
              type: string
            displayScript:
              type: string
            displayCountry:
              type: string
            displayVariant:
              type: string
            displayName:
              type: string
        cacheControl:
          type: string
        basicAuth:
          type: string
          writeOnly: true
        etag:
          type: string
        accept:
          type: array
          items:
            $ref: '#/components/schemas/MediaType'
        acceptLanguage:
          type: array
          items:
            type: object
            properties:
              range:
                type: string
              weight:
                type: number
                format: double
        acceptLanguageAsLocales:
          type: array
          items:
            type: object
            properties:
              language:
                type: string
              script:
                type: string
              country:
                type: string
              variant:
                type: string
              extensionKeys:
                uniqueItems: true
                type: array
                items:
                  type: string
              unicodeLocaleAttributes:
                uniqueItems: true
                type: array
                items:
                  type: string
              unicodeLocaleKeys:
                uniqueItems: true
                type: array
                items:
                  type: string
              iso3Language:
                type: string
              iso3Country:
                type: string
              displayLanguage:
                type: string
              displayScript:
                type: string
              displayCountry:
                type: string
              displayVariant:
                type: string
              displayName:
                type: string
        accessControlAllowCredentials:
          type: boolean
        accessControlAllowHeaders:
          type: array
          items:
            type: string
        accessControlAllowMethods:
          type: array
          items:
            type: string
            enum:
            - GET
            - HEAD
            - POST
            - PUT
            - PATCH
            - DELETE
            - OPTIONS
            - TRACE
        accessControlAllowOrigin:
          type: string
        accessControlExposeHeaders:
          type: array
          items:
            type: string
        accessControlMaxAge:
          type: integer
          format: int64
        accessControlRequestHeaders:
          type: array
          items:
            type: string
        accessControlRequestMethod:
          type: string
          enum:
          - GET
          - HEAD
          - POST
          - PUT
          - PATCH
          - DELETE
          - OPTIONS
          - TRACE
        bearerAuth:
          type: string
          writeOnly: true
        contentDisposition:
          $ref: '#/components/schemas/ContentDisposition'
        expires:
          type: integer
          format: int64
        ifMatch:
          type: array
          items:
            type: string
        ifNoneMatch:
          type: array
          items:
            type: string
        ifUnmodifiedSince:
          type: integer
          format: int64
        pragma:
          type: string
        upgrade:
          type: string
        vary:
          type: array
          items:
            type: string
      additionalProperties:
        type: array
        items:
          type: string
    HttpRange:
      type: object
    MediaType:
      type: object
      properties:
        type:
          type: string
        subtype:
          type: string
        parameters:
          type: object
          additionalProperties:
            type: string
        qualityValue:
          type: number
          format: double
        charset:
          type: object
          properties:
            registered:
              type: boolean
        wildcardType:
          type: boolean
        wildcardSubtype:
          type: boolean
        concrete:
          type: boolean
    ResponseEntityObject:
      type: object
      properties:
        headers:
          type: object
          properties:
            empty:
              type: boolean
            location:
              type: string
              format: uri
            lastModified:
              type: integer
              format: int64
            host:
              type: object
              properties:
                address:
                  type: object
                  properties:
                    canonicalHostName:
                      type: string
                    address:
                      type: array
                      items:
                        type: string
                        format: byte
                    hostAddress:
                      type: string
                    hostName:
                      type: string
                    multicastAddress:
                      type: boolean
                    anyLocalAddress:
                      type: boolean
                    loopbackAddress:
                      type: boolean
                    linkLocalAddress:
                      type: boolean
                    siteLocalAddress:
                      type: boolean
                    mcglobal:
                      type: boolean
                    mcnodeLocal:
                      type: boolean
                    mclinkLocal:
                      type: boolean
                    mcsiteLocal:
                      type: boolean
                    mcorgLocal:
                      type: boolean
                port:
                  type: integer
                  format: int32
                unresolved:
                  type: boolean
                hostName:
                  type: string
                hostString:
                  type: string
            all:
              type: object
              additionalProperties:
                type: string
              writeOnly: true
            range:
              type: array
              items:
                $ref: '#/components/schemas/HttpRange'
            contentLength:
              type: integer
              format: int64
            contentType:
              $ref: '#/components/schemas/MediaType'
            date:
              type: integer
              format: int64
            ifModifiedSince:
              type: integer
              format: int64
            connection:
              type: array
              items:
                type: string
            origin:
              type: string
            acceptCharset:
              type: array
              items:
                type: object
                properties:
                  registered:
                    type: boolean
            allow:
              uniqueItems: true
              type: array
              items:
                type: string
                enum:
                - GET
                - HEAD
                - POST
                - PUT
                - PATCH
                - DELETE
                - OPTIONS
                - TRACE
            contentLanguage:
              type: object
              properties:
                language:
                  type: string
                script:
                  type: string
                country:
                  type: string
                variant:
                  type: string
                extensionKeys:
                  uniqueItems: true
                  type: array
                  items:
                    type: string
                unicodeLocaleAttributes:
                  uniqueItems: true
                  type: array
                  items:
                    type: string
                unicodeLocaleKeys:
                  uniqueItems: true
                  type: array
                  items:
                    type: string
                iso3Language:
                  type: string
                iso3Country:
                  type: string
                displayLanguage:
                  type: string
                displayScript:
                  type: string
                displayCountry:
                  type: string
                displayVariant:
                  type: string
                displayName:
                  type: string
            cacheControl:
              type: string
            basicAuth:
              type: string
              writeOnly: true
            etag:
              type: string
            accept:
              type: array
              items:
                $ref: '#/components/schemas/MediaType'
            acceptLanguage:
              type: array
              items:
                type: object
                properties:
                  range:
                    type: string
                  weight:
                    type: number
                    format: double
            acceptLanguageAsLocales:
              type: array
              items:
                type: object
                properties:
                  language:
                    type: string
                  script:
                    type: string
                  country:
                    type: string
                  variant:
                    type: string
                  extensionKeys:
                    uniqueItems: true
                    type: array
                    items:
                      type: string
                  unicodeLocaleAttributes:
                    uniqueItems: true
                    type: array
                    items:
                      type: string
                  unicodeLocaleKeys:
                    uniqueItems: true
                    type: array
                    items:
                      type: string
                  iso3Language:
                    type: string
                  iso3Country:
                    type: string
                  displayLanguage:
                    type: string
                  displayScript:
                    type: string
                  displayCountry:
                    type: string
                  displayVariant:
                    type: string
                  displayName:
                    type: string
            accessControlAllowCredentials:
              type: boolean
            accessControlAllowHeaders:
              type: array
              items:
                type: string
            accessControlAllowMethods:
              type: array
              items:
                type: string
                enum:
                - GET
                - HEAD
                - POST
                - PUT
                - PATCH
                - DELETE
                - OPTIONS
                - TRACE
            accessControlAllowOrigin:
              type: string
            accessControlExposeHeaders:
              type: array
              items:
                type: string
            accessControlMaxAge:
              type: integer
              format: int64
            accessControlRequestHeaders:
              type: array
              items:
                type: string
            accessControlRequestMethod:
              type: string
              enum:
              - GET
              - HEAD
              - POST
              - PUT
              - PATCH
              - DELETE
              - OPTIONS
              - TRACE
            bearerAuth:
              type: string
              writeOnly: true
            contentDisposition:
              $ref: '#/components/schemas/ContentDisposition'
            expires:
              type: integer
              format: int64
            ifMatch:
              type: array
              items:
                type: string
            ifNoneMatch:
              type: array
              items:
                type: string
            ifUnmodifiedSince:
              type: integer
              format: int64
            pragma:
              type: string
            upgrade:
              type: string
            vary:
              type: array
              items:
                type: string
          additionalProperties:
            type: array
            items:
              type: string
        body:
          type: object
        statusCode:
          type: string
          enum:
          - 100 CONTINUE
          - 101 SWITCHING_PROTOCOLS
          - 102 PROCESSING
          - 103 CHECKPOINT
          - 200 OK
          - 201 CREATED
          - 202 ACCEPTED
          - 203 NON_AUTHORITATIVE_INFORMATION
          - 204 NO_CONTENT
          - 205 RESET_CONTENT
          - 206 PARTIAL_CONTENT
          - 207 MULTI_STATUS
          - 208 ALREADY_REPORTED
          - 226 IM_USED
          - 300 MULTIPLE_CHOICES
          - 301 MOVED_PERMANENTLY
          - 302 FOUND
          - 302 MOVED_TEMPORARILY
          - 303 SEE_OTHER
          - 304 NOT_MODIFIED
          - 305 USE_PROXY
          - 307 TEMPORARY_REDIRECT
          - 308 PERMANENT_REDIRECT
          - 400 BAD_REQUEST
          - 401 UNAUTHORIZED
          - 402 PAYMENT_REQUIRED
          - 403 FORBIDDEN
          - 404 NOT_FOUND
          - 405 METHOD_NOT_ALLOWED
          - 406 NOT_ACCEPTABLE
          - 407 PROXY_AUTHENTICATION_REQUIRED
          - 408 REQUEST_TIMEOUT
          - 409 CONFLICT
          - 410 GONE
          - 411 LENGTH_REQUIRED
          - 412 PRECONDITION_FAILED
          - 413 PAYLOAD_TOO_LARGE
          - 413 REQUEST_ENTITY_TOO_LARGE
          - 414 URI_TOO_LONG
          - 414 REQUEST_URI_TOO_LONG
          - 415 UNSUPPORTED_MEDIA_TYPE
          - 416 REQUESTED_RANGE_NOT_SATISFIABLE
          - 417 EXPECTATION_FAILED
          - 418 I_AM_A_TEAPOT
          - 419 INSUFFICIENT_SPACE_ON_RESOURCE
          - 420 METHOD_FAILURE
          - 421 DESTINATION_LOCKED
          - 422 UNPROCESSABLE_ENTITY
          - 423 LOCKED
          - 424 FAILED_DEPENDENCY
          - 425 TOO_EARLY
          - 426 UPGRADE_REQUIRED
          - 428 PRECONDITION_REQUIRED
          - 429 TOO_MANY_REQUESTS
          - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
          - 451 UNAVAILABLE_FOR_LEGAL_REASONS
          - 500 INTERNAL_SERVER_ERROR
          - 501 NOT_IMPLEMENTED
          - 502 BAD_GATEWAY
          - 503 SERVICE_UNAVAILABLE
          - 504 GATEWAY_TIMEOUT
          - 505 HTTP_VERSION_NOT_SUPPORTED
          - 506 VARIANT_ALSO_NEGOTIATES
          - 507 INSUFFICIENT_STORAGE
          - 508 LOOP_DETECTED
          - 509 BANDWIDTH_LIMIT_EXCEEDED
          - 510 NOT_EXTENDED
          - 511 NETWORK_AUTHENTICATION_REQUIRED
        statusCodeValue:
          type: integer
          format: int32
    ExerciseMappingHelper:
      required:
      - calories
      - duration
      type: object
      properties:
        description:
          type: string
          enum:
          - RUNNING
          - SWIMMING
          - STRENGTH_TRAINING
          - CIRCUT_TRAINING
        startTime:
          type: string
          format: date-time
        duration:
          type: integer
          format: int32
        calories:
          type: integer
          format: int32
        userId:
          type: integer
          format: int32
        id:
          type: integer
          format: int32
    UserA:
      type: object
      properties:
        userId:
          type: integer
          format: int32
        points:
          type: number
          format: float
    UserMappingHelper:
      type: object
      properties:
        userId:
          type: integer
          format: int32
        points:
          type: number
          format: float
        usera:
          $ref: '#/components/schemas/UserA'
        id:
          type: integer
          format: int32